% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevance_model.R
\name{embed_text}
\alias{embed_text}
\title{Embed a character vector with a planscanR embedding model.}
\usage{
embed_text(model, x)
}
\arguments{
\item{model}{A \code{planscanR_embedding_model} object.}

\item{x}{Character vector to embed.}
}
\value{
Numeric matrix.
}
\description{
Returns a numeric matrix with one row per input element.
}
