% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevance_model.R
\name{model_name}
\alias{model_name}
\title{Identify a model with a stable name string for logging / sidecars.}
\usage{
model_name(model)
}
\arguments{
\item{model}{A \code{planscanR_embedding_model} object.}
}
\value{
Character scalar.
}
\description{
Identify a model with a stable name string for logging / sidecars.
}
